refactor(views): 뷰 폴더로 이동
