+++ src/views/OutputView.js
@@ -1,5 +1,5 @@
 const { Console } = require('@woowacourse/mission-utils');
-const { BRIDGE_STRING } = require('./constants');
+const { BRIDGE_STRING } = require('../constants');
 
 const OutputView = {
   START_MESSAGE: '다리 건너기 게임을 시작합니다.',
